backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
billiard==4.2.4
CacheControl==0.14.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
fastapi==0.115.6
firebase_admin==7.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.26.0
google-cloud-storage==3.4.1
google-crc32c==1.8.0
google-generativeai==0.8.4
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.1.1
kombu==5.6.2
langchain-core==1.2.23
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
msgpack==1.1.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
redis==7.4.0
requests==2.32.5
requests-toolbelt==1.0.0
resend==2.26.0
six==1.17.0
starlette==0.41.3
tenacity==9.1.4
tqdm==4.67.3
tree-sitter==0.25.2
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.6.0
xxhash==3.6.0
zstandard==0.25.0
